fix undefined in Check
